+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## 1.0.2 (2024-02-23)
 
+- Automatically mark context as dirty if `#safe_rewrite` modifies source code. ([@palkan][])
+
 - Add `context.path` to rewriters to access the current file path (if any). ([@palkan][])
 
 ## 1.0.1 (2024-01-28)